[BE] refactor: 카테고리 내 히어릿 목록 찾기의 도메인을 Category에서 Hearit으로 변경 #227
