+++ backend/src/test/java/com/onair/hearit/presentation/HearitControllerTest.java
@@ -235,6 +235,40 @@ class HearitControllerTest extends IntegrationTest {
         );
     }
 
+    @Test
+    @DisplayName("카테고리로 히어릿 검색 시 200 OK 및 해당 카테고리의 히어릿들을 최신순으로 반환한다.")
+    void searchHearitsByCategoryWithPagination() {
+        // given
+        Category category1 = dbHelper.insertCategory(new Category("Spring", "#001"));
+        Category category2 = dbHelper.insertCategory(new Category("Java", "#002"));
+
+        Hearit hearit1 = dbHelper.insertHearit(TestFixture.createFixedHearitWith(category1));
+        Hearit hearit2 = dbHelper.insertHearit(TestFixture.createFixedHearitWith(category1));
+        dbHelper.insertHearit(TestFixture.createFixedHearitWith(category2)); // 다른 카테고리
+
+        // when
+        PagedResponse<HearitSearchResponse> pagedResponse = RestAssured
+                .given()
+                .pathParam("categoryId", category1.getId())
+                .queryParam("page", 0)
+                .queryParam("size", 10)
+                .when()
+                .get("/api/v1/categories/{categoryId}/hearits")
+                .then()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .as(new TypeRef<>() {
+                });
+        List<HearitSearchResponse> responses = pagedResponse.content();
+
+        // then
+        assertAll(
+                () -> assertThat(responses).hasSize(2),
+                () -> assertThat(responses.get(0).id()).isEqualTo(hearit2.getId()), // 최신 hearit 먼저
+                () -> assertThat(responses.get(1).id()).isEqualTo(hearit1.getId())
+        );
+    }
+
     private String generateToken(Member member) {
         return jwtTokenProvider.createToken(member.getId());
     }
